# user-manager/user-manager-rest-api/openapi/user-manager-rest-api.yml
...
openapi: 3.0.0
info:

  # arbitrary string that specifies the version of the API
  version: 0.0.1

  # API name
  title: Butterfly User Manager API

  # Extended information about the API
  description: >-
    Butterfly User Manager REST API

  contact:
    name: DStack Group
    email: dstackgroup@gmail.com
    # url: https://example.com/

# API server and base URL
servers:
  - url: http://localhost:5000
    description: Development Server

# tags are used for organizing operations
...
paths:
...
  /users:
    get:
      operationId: getUsers
      summary: Retrieves the list of registered users
      tags: [ 'USERS' ]
      responses:
        '200':
          description: OK, returns a list of user records.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'

    post:
      operationId: createUser
      summary: Creates new user and returns the user record
      tags: [ 'USERS' ]
      requestBody:
        description: User record
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUser'
      responses:
        '201':
          description: OK, user inserted
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/User'
        '409':
          description: Unique constraint violation, another user exists with the same email
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    description: error message
  /users/{email}:
    get:
      operationId: getUserByEmail
      summary: Searches a user by its email and returns its record data
      tags: [ 'USERS' ]
      parameters:
        - in: path
          name: email
          schema:
            type: string
            format: email
          required: true
      responses:
        '200':
          description: OK, user found
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/User'
        '404':
          description: Error, the user does not exist
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    description: error message

    patch:
      operationId: patchUserByEmail
      summary: Updates the data of the user identified by the given email and returns the updated user record
      tags: [ 'USERS' ]
      parameters:
      - in: path
        name: email
        schema:
          type: string
          format: email
        required: true
      requestBody:
        description: User record
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUser'
      responses:
        '200':
          description: OK, user updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/User'
        '404':
          description: Error, the user does not exist
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    description: error message

    delete:
      operationId: deleteUserByEmail
      summary: Removes the user that is registered with the given email
      tags: [ 'USERS' ]
      parameters:
        - in: path
          name: email
          schema:
            type: string
            format: email
          required: true
      responses:
        '200':
          description: OK, user deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    nullable: true
        '404':
          description: Error, the user does not exist
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    description: error message
...
components:
  schemas:
...
    CreateUser:
      type: object
      required:
        - email
        - firstname
        - lastname
      properties:
        email:
          type: string
        firstname:
          type: string
        lastname:
          type: string
        enabled:
          type: boolean
          default: true
      example:
        email: "dstackgroup@gmail.com"
        firstname: "DStack"
        lastname: "Group"
    UpdateUser:
      type: object
      required:
        - email
      properties:
        email:
          type: string
        firstname:
          type: string
        lastname:
          type: string
        enabled:
          type: boolean
          default: true
      example:
        email: "dstackgroup@gmail.com"
        enabled: false
    User:
      type: object
      required:
        - userId
        - email
        - firstname
        - lastname
        - created
      properties:
        userId:
          type: string
        email:
          type: string
        firstname:
          type: string
        lastname:
          type: string
        enabled:
          type: boolean
        created:
          type: string
          format: date-time
        modified:
          type: string
          format: date-time
          nullable: true
      example:
        userId: "1"
        email: "dstackgroup@gmail.com"
        firstname: "DStack"
        lastname: "Group"
        enabled: true
...